+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Modify
 - avoiding check
+- figure color to short name
 
 ## [0.3.3] - 2019-01-17
 ### Added